refactor: 移除动态文本代码，使用文本元素替代
