+++ packages/effects-core/package.json
@@ -50,7 +50,7 @@
     "registry": "https://registry.npmjs.org"
   },
   "dependencies": {
-    "@galacean/effects-specification": "2.0.0-alpha.2",
+    "@galacean/effects-specification": "link://Users/yiqi/git/galacean/effects-specification",
     "@galacean/effects-math": "1.0.1",
     "uuid": "9.0.1"
   }
